partial generalblock reorg
