+++ avallama.Tests/ObservableStackTests.cs
@@ -1,3 +1,6 @@
+// Copyright (c) Márk Csörgő and Martin Bartos
+// Licensed under the MIT License. See LICENSE file for details.
+
 using System;
 using avallama.Utilities;
 using Xunit;
