+++ avallama/Styles/TextItem.axaml
@@ -1,4 +1,4 @@
-﻿<!-- 
+﻿<!--
 Copyright (c) Márk Csörgő and Martin Bartos
 Licensed under the MIT License. See LICENSE file for details.
 -->
@@ -11,17 +11,16 @@ Licensed under the MIT License. See LICENSE file for details.
         <ThemeVariantScope RequestedThemeVariant="Light">
             <Border Padding="20" Background="{DynamicResource Surface}">
                 <StackPanel Orientation="Vertical" Spacing="10">
-                    <controls:GeneralBlock
+                    <controls:MessageItem
                         Classes="userMessage"
                         Text="User Message"/>
-                
-                    <controls:GeneralBlock 
+
+                    <controls:MessageItem
                         Classes="generatedMessage"
                         Text="Generated Message"
                         SubText="SubText"/>
-                    
-                    <controls:GeneralBlock
-                        Classes="conversationElement"
+
+                    <controls:ConversationItem
                         Text="Conversation Title"
                         Width="300"
                         SubText="llama3.2"/>
@@ -29,9 +28,8 @@ Licensed under the MIT License. See LICENSE file for details.
             </Border>
         </ThemeVariantScope>
     </Design.PreviewWith>
-    
-    <Style Selector="controls|GeneralBlock">
-        <Setter Property="Clickable" Value="False"/>
+
+    <Style Selector="controls|TextItem, controls|MessageItem, controls|ConversationItem">
         <Setter Property="TextAlignment" Value="Left"/>
         <Setter Property="TextFontSize" Value="14"/>
         <Setter Property="CornerRadius" Value="20"/>
@@ -40,30 +38,26 @@ Licensed under the MIT License. See LICENSE file for details.
         <Setter Property="LineHeight" Value="22"/>
         <Setter Property="FontFamily" Value="{StaticResource Manrope}"/>
     </Style>
-    
-    <Style Selector="controls|GeneralBlock.userMessage">
-        <Setter Property="Selectable" Value="True"/>
+
+    <Style Selector="controls|MessageItem.userMessage">
         <Setter Property="Background" Value="{DynamicResource Primary}"/>
         <Setter Property="TextColor" Value="{DynamicResource OnPrimary}"/>
         <Setter Property="HorizontalAlignment" Value="Right"/>
-        <Setter Property="SelectionColor" Value="{DynamicResource OnPrimary}"/>
+        <Setter Property="SelectionBrush" Value="{DynamicResource OnPrimary}"/>
     </Style>
-    
-    <Style Selector="controls|GeneralBlock.generatedMessage">
-        <Setter Property="Selectable" Value="True"/>
+
+    <Style Selector="controls|MessageItem.generatedMessage">
         <Setter Property="Background" Value="{DynamicResource SurfaceContainerHighest}"/>
         <Setter Property="TextColor" Value="{DynamicResource OnSurfaceVariant}"/>
         <Setter Property="SubTextAlignment" Value="End"/>
         <Setter Property="HorizontalAlignment" Value="Left"/>
         <Setter Property="SubTextFontSize" Value="10"/>
         <Setter Property="SubTextColor" Value="{DynamicResource Outline}"/>
-        <Setter Property="SelectionColor" Value="{DynamicResource Primary}"/>
+        <Setter Property="SelectionBrush" Value="{DynamicResource Primary}"/>
         <Setter Property="Spacing" Value="5"/>
     </Style>
-    
-    <Style Selector="controls|GeneralBlock.conversationElement">
-        <Setter Property="Selectable" Value="False"/>
-        <Setter Property="Clickable" Value="True"/>
+
+    <Style Selector="controls|ConversationItem">
         <Setter Property="Background" Value="{DynamicResource SurfaceContainerHighest}"/>
         <Setter Property="TextColor" Value="{DynamicResource OnSurface}"/>
         <Setter Property="TextFontSize" Value="16"/>
