+++ avallama/avallama.csproj
@@ -1,4 +1,4 @@
-﻿<!-- 
+﻿<!--
 Copyright (c) Márk Csörgő and Martin Bartos
 Licensed under the MIT License. See LICENSE file for details.
 -->
@@ -33,14 +33,14 @@ Licensed under the MIT License. See LICENSE file for details.
         <PackageReference Include="Avalonia.Desktop" Version="11.3.6" />
         <PackageReference Include="Avalonia.Themes.Fluent" Version="11.3.6" />
         <PackageReference Include="Avalonia.Fonts.Inter" Version="11.3.6" />
-        
+
         <!--Condition below is needed to remove Avalonia.Diagnostics package from build output in Release configuration.-->
         <PackageReference Include="Avalonia.Diagnostics" Version="11.3.6">
             <IncludeAssets Condition="'$(Configuration)' != 'Debug'">None</IncludeAssets>
             <PrivateAssets Condition="'$(Configuration)' != 'Debug'">All</PrivateAssets>
         </PackageReference>
         <PackageReference Include="HtmlAgilityPack" Version="1.12.3" />
-        
+
         <PackageReference Include="Svg.Controls.Avalonia" Version="11.3.6.2" />
         <PackageReference Include="CommunityToolkit.Mvvm" Version="8.4.0" />
         <PackageReference Include="Microsoft.Data.Sqlite" Version="9.0.9" />
@@ -62,5 +62,5 @@ Licensed under the MIT License. See LICENSE file for details.
     <ItemGroup>
       <Folder Include="Assets\Images\" />
     </ItemGroup>
-    
+
 </Project>
